% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installing.rcasc.R
\name{installing.rcasc}
\alias{installing.rcasc}
\title{A function prepare the environment for the SCA tutorial.}
\usage{
installing.rcasc()
}
\description{
This function check that dcker is installed and download locally the required docker containers
}
\examples{
\dontrun{

installing.rcasc()

}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy
}
