\name{SCAtutorial-package}
\alias{SCAtutorial-package}
\alias{SCAtutorial}
\docType{package}
\title{
\packageTitle{SCAtutorial}
}
\description{
\packageDescription{SCAtutorial}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SCAtutorial}
\packageIndices{SCAtutorial}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SCAtutorial}

Maintainer: \packageMaintainer{SCAtutorial}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
